+++ CHANGELOG.md
@@ -13,7 +13,10 @@
     - **Bugfix:** Adjusted `UnmarshalJSON` function to use enum types and added tests for enums
 - `git`: [v0.5.1](services/git/CHANGELOG.md#v051-2025-06-04)
     - **Bugfix:** Adjusted `UnmarshalJSON` function to use enum types and added tests for enums
-- `loadbalancer`: [v1.2.1](services/loadbalancer/CHANGELOG.md#v121-2025-06-04)
+- `loadbalancer`: 
+  - [v1.3.0](services/loadbalancer/CHANGELOG.md#v130-2025-06-10)
+    - **Feature:** Add new field `TargetSecurityGroup` in `LoadBalancer` Model
+  - [v1.2.1](services/loadbalancer/CHANGELOG.md#v121-2025-06-04)
     - **Bugfix:** Adjusted `UnmarshalJSON` function to use enum types and added tests for enums
 - `logme`: [v0.24.1](services/logme/CHANGELOG.md#v0241-2025-06-04)
     - **Bugfix:** Adjusted `UnmarshalJSON` function to use enum types and added tests for enums